aded comments

+++ app/Http/Controllers/Comment/DestroyController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers\Comment;
 
 use App\Models\Comment;
+use App\Models\Post;
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 
@@ -16,10 +17,10 @@ class DestroyController extends Controller
    */
   public function __invoke(Request $request, Comment $comment)
   {
-    // if ($request->user()->cannot('delete', $comment)) {
-    //   abort(403);
-    // }
+    if ($request->user()->cannot('delete', $comment)) {
+      abort(403);
+    }
     $comment->destroy($comment->id);
-    return redirect()->route('posts.index');
+    return redirect()->route('posts.show', $comment->post);
   }
 }
